Drone changes :
Changed the value of the navigation to lower ones in order to provide a more precise way of control.
Added the HoverMode in the key binding to be able to stabilize the drone in case of emergency.
Added the possibility of changing the camera of the drone. --> camera from under the drone freeze the display.